# docs/swagger.yaml
definitions:
  jsonpatch.Operation:
    additionalProperties:
      items:
        type: integer
      type: array
    type: object
  pb.AT_DRONE_STATUS:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - AT_DRONE_STATUS_ADS_UNKNOWN
    - AT_DRONE_STATUS_ADS_FLYING
    - AT_DRONE_STATUS_ADS_LANDED
  pb.AT_VALIDATE_STATUS:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - AT_VALIDATE_STATUS_ADS_CREATED
    - AT_VALIDATE_STATUS_ADS_SENT
    - AT_VALIDATE_STATUS_ADS_APPROVED
...
  pb.Drone:
    properties:
      created_at:
        type: integer
      created_by:
        type: string
      deport_id:
        type: string
      drone_status:
        $ref: '#/definitions/pb.AT_DRONE_STATUS'
      fixed_locker_ids:
        items:
          type: string
        type: array
      gcs_id:
        type: string
      id:
        type: string
      image:
        type: string
      mass:
        type: number
      max_battery:
        type: number
      max_payload:
        type: number
      max_power:
        type: number
      max_speed:
        type: number
      rtsp_link:
        type: string
      size:
        items:
          type: number
        type: array
      updated_at:
        type: integer
      updated_by:
        type: string
      validate_status:
        $ref: '#/definitions/pb.AT_VALIDATE_STATUS'
    type: object
...
  pb.PatchResponse:
    properties:
      IsOk:
        type: boolean
      Message:
        type: string
    type: object
...
  types.ErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample Go Template server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Go Template Project API
  version: "1.0"
paths:
  /drones:
    get:
      consumes:
      - application/json
      description: Find all drone
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.Drone'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Find all drone
      tags:
      - drones
    post:
      consumes:
      - application/json
      description: Create a new drone
      parameters:
      - description: drone body
        in: body
        name: drone
        required: true
        schema:
          $ref: '#/definitions/pb.Drone'
      - description: event api call flag
        in: query
        name: eventAPI
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.Drone'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Create a drone
      tags:
      - drones
  /drones/{id}:
    delete:
      consumes:
      - application/json
      description: Delete drone by ID
      parameters:
      - description: drone id
        in: path
        name: id
        required: true
        type: string
      - description: event api call flag
        in: query
        name: eventAPI
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.Drone'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Delete drone by ID
      tags:
      - drones
    get:
      consumes:
      - application/json
      description: Find drone by ID
      parameters:
      - description: drone id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.Drone'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Find drone by ID
      tags:
      - drones
    patch:
      consumes:
      - application/json
      description: Patch drone by ID use standard of JSON PATCH https://jsonpatch.com/
      parameters:
      - description: drone id
        in: path
        name: id
        required: true
        type: string
      - description: Patch operation format Array of Operation Add, Remove, Replace,
          Copy, Move, Test. Get example at https://jsonpatch.com/
        in: body
        name: drone
        required: true
        schema:
          items:
            $ref: '#/definitions/jsonpatch.Operation'
          type: array
      - description: event api call flag
        in: query
        name: eventAPI
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.PatchResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Patch drone by ID
      tags:
      - drones
    put:
      consumes:
      - application/json
      description: Update drone by ID
      parameters:
      - description: drone id
        in: path
        name: id
        required: true
        type: string
      - description: drone body
        in: body
        name: drone
        required: true
        schema:
          $ref: '#/definitions/pb.Drone'
      - description: event api call flag
        in: query
        name: eventAPI
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.Drone'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Update drone by ID
      tags:
      - drones
  /drones/search:
    get:
      consumes:
      - application/json
      description: Search drone use Query option https://github.com/jtlabsio/mongo/
      parameters:
      - description: page number
        in: query
        name: page[page]
        required: true
        type: integer
      - description: page size
        in: query
        name: page[size]
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.Drone'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Search drone
      tags:
      - drones
...
swagger: "2.0"